Extracted Package and Location from the Shipping class.
Cuz single responsibility and all that jazz.
