+++ app/models/shipping_location.rb
@@ -1,4 +1,8 @@
 require 'active_shipping'
 
-class ShippingLocation
+class ShippingLocation < ActiveShipping::Location
+  include ActiveModel::Validations
+
+  validates :country, :city, :province, :postal_code, presence: true
+
 end
