+++ spec/models/shipping_rate_spec.rb
@@ -1,0 +1,6 @@
+require 'rails_helper'
+require 'shipping_rate'
+
+RSpec.describe ShippingRate do
+  pending "validations"
+end
